% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDGM3.R
\name{SDGM3}
\alias{SDGM3}
\title{Survival Data Generation Method 3}
\usage{
SDGM3(N = 200, p = 15, u_max = 7)
}
\arguments{
\item{N}{The sample size of the simulated dataset.}

\item{p}{The covariate dimension of the simulated dataset.}

\item{u_max}{The parameter which is used to control the censoring rate.}
}
\value{
the simulated dataset
}
\description{
Survival data generation method. The proportional hazards assumption is strongly violated in this dataset.
}
\examples{
SDGM3(N = 200, p = 15, u_max = 7)

}
\references{
Steingrimsson, J. A. ,  Diao, L. , &  Strawderman, R. L. . (2019). Censoring unbiased regression trees and ensembles. Journal of the American Statistical Association, 114.

Zhu, R. , &  Kosorok, M. R. . (2012). Recursively imputed survival trees. Journal of the American Statistical Association, 107(497), 331-340.

Ishwaran, H. ,  Kogalur, U. B. ,  Gorodeski, E. Z. ,  Minn, A. J. , &  Lauer, M. S. . (2010). High-dimensional variable selection for survival data. Journal of the American Statistical Association, 105(489), 205-217.
}
\author{
Hanpu Zhou \email{zhouhanpu@csu.edu.cn}
}
