% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvMetrics-package.R
\docType{package}
\name{SurvMetrics-package}
\alias{SurvMetrics}
\alias{SurvMetrics-package}
\title{SurvMetrics: Predictive Evaluation Metrics in Survival Analysis}
\description{
An implementation of popular evaluation metrics that are commonly used in survival prediction including Concordance Index, Brier Score, Integrated Brier Score, Integrated Square Error, Integrated Absolute Error and Mean Absolute Error. For a detailed information, see (Ishwaran H, Kogalur UB, Blackstone EH and Lauer MS (2008) \doi{10.1214/08-AOAS169}) , (Moradian H, Larocque D and Bellavance F (2017) \doi{10.1007/s10985-016-9372-1}), (Hanpu Zhou, Hong Wang, Sizheng Wang and Yi Zou (2023) \doi{10.32614/rj-2023-009}) for different evaluation metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whcsu/SurvMetrics}
}

}
\author{
\strong{Maintainer}: Hong Wang \email{wh@csu.edu.cn}

Authors:
\itemize{
  \item Hanpu Zhou \email{974637739@qq.com}
  \item Xuewei Cheng \email{xwcheng@hunnu.edu.cn}
  \item Sizheng Wang \email{964555106@qq.com}
  \item Yi Zou \email{2584695843@qq.com}
}

}
\keyword{internal}
