% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBS.R
\name{IBS}
\alias{IBS}
\title{IBS}
\usage{
IBS(object, sp_matrix, IBSrange = range(object[, 1]))
}
\arguments{
\item{object}{object of class \code{Surv} in the testing set created by Surv function.}

\item{sp_matrix}{a matrix or data.frame of predicted values of survival probabilities for the testing set.}

\item{IBSrange}{a vector contains all discrete time points corresponding to the predicted probability in sp_matrix.
Or the scale you want to get the IBS; and if it is a single point the return value will be the Brier Score at the timepoint.}
}
\value{
The integration of brierscore
}
\description{
IBS is  an integrated version of the Brier which is used to calculate the integration of the Brier Score.
The Brier Score is the mean square difference between the true classes and the predicted probabilities.
Basically, the IBS is an integrated
weighted squared distance between the estimated survival function and the empirical
survival function. The inverse probability censoring weighting(IPCW) is used to adjust for censoring.
}
\details{
The percentage of censored observations increases in time, and this will surely affect the dispersion of the empirical Brier Score.
The question of how censoring in finite samples acts on the distribution of our measures of inaccuracy is an interesting subject.
Our recommendation is to choose t* in a way that censoring is not too heavy (for example, the median follow-up time).
We also prefer measures with integrated loss functions since they will reflect inaccuracy over an interval rather than just at one point in time.
In addition, the corresponding empirical measures are likely to have lower dispersion, because censored observations contribute their estimated event-free probabilities to the integrand until the censoring occurs.
}
\examples{
library(randomForestSRC)
library(MASS)
library(survival)
set.seed(123)
 N=100
 p = 25
 c.mean = 0.4
 mu = rep(0,p)
 Si = matrix(0,p,p)
 for (i in 1:p) {
   for (j in 1:p) {
     Si[i,j] = 0.9^abs(i-j)
   }
 }
 W = mvrnorm(N,mu,Si)
 Ti = rep(0,N)
 for (i in 1:N) {
   t.mu = exp(0.1*sum(W[i,1:round(p/2)]))
   Ti[i] = rexp(1,1/t.mu)
 }
 c.time = rexp(N,c.mean)
 mydata.x = as.data.frame(W)
 mydata.time = data.frame('time'=Ti,'c.time' = c.time,'status' = 0)
 mydata.time$status = ifelse(mydata.time$time < mydata.time$c.time,1,0)
 mydata.time$time = apply(mydata.time[,1:2],1,min)
 mydata = data.frame('time' = mydata.time$time,'status' = mydata.time$status,W)

index.train = sample(1:N,2/3*N)
data.train = mydata[index.train,]
data.test = mydata[-index.train,]

fit.RSF = rfsrc(Surv(time,status)~.,data.train,nsplit=3,ntree=500)
predicted = predict(fit.RSF,data.test)
sp_matrix = predicted$survival

object = Surv(data.test$time,data.test$status)

#the default time points
IBS(object,sp_matrix)
#a time range
IBS(object,sp_matrix,c(18:100))

}
\references{
HooraMoradian, DenisLarocque, & FranoisBellavance. (2017). \\(l_1\\) splitting rules in survival forests. Lifetime Data Analysis, 23(4), 671–691.

Graf, Erika, Schmoor, Claudia, Sauerbrei, & Willi, et al. (1999). Assessment and comparison of prognostic classification schemes for survival data. Statist. Med., 18(1718), 2529-2545.

Brier, G. W. . (1950). Verification of forecasts expressed in terms of probability. Monthly Weather Review, 78.

Gneiting, T. , &  Raftery, A. E. . (2007). Strictly Proper Scoring Rules, Prediction, and Estimation.
}
\author{
Zhou HanPu \email{zhouhanpu@csu.edu.cn}
}
